Split WebSocket actions and events into multiple files
