+++ app/actions/websocket/integrations.ts
@@ -1,0 +1,14 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import {IntegrationTypes} from '@mm-redux/action_types';
+import {ActionResult, DispatchFunc} from '@mm-redux/types/actions';
+import {WebSocketMessage} from '@mm-redux/types/websocket';
+
+export function handleOpenDialogEvent(msg: WebSocketMessage) {
+    return (dispatch: DispatchFunc): ActionResult => {
+        const data = (msg.data && msg.data.dialog) || {};
+        dispatch({type: IntegrationTypes.RECEIVED_DIALOG, data: JSON.parse(data)});
+        return {data: true};
+    };
+}
