+++ app/actions/websocket/roles.ts
@@ -1,0 +1,33 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import {RoleTypes} from '@mm-redux/action_types';
+import {GenericAction} from '@mm-redux/types/actions';
+import {WebSocketMessage} from '@mm-redux/types/websocket';
+
+export function handleRoleAddedEvent(msg: WebSocketMessage): GenericAction {
+    const role = JSON.parse(msg.data.role);
+
+    return {
+        type: RoleTypes.RECEIVED_ROLE,
+        data: role,
+    };
+}
+
+export function handleRoleRemovedEvent(msg: WebSocketMessage): GenericAction {
+    const role = JSON.parse(msg.data.role);
+
+    return {
+        type: RoleTypes.ROLE_DELETED,
+        data: role,
+    };
+}
+
+export function handleRoleUpdatedEvent(msg: WebSocketMessage): GenericAction {
+    const role = JSON.parse(msg.data.role);
+
+    return {
+        type: RoleTypes.RECEIVED_ROLE,
+        data: role,
+    };
+}
